Renamed plugin compatibility modules. PluginIntegration, Plugins, and CompatibilityManager are now interfaces to improve the readability.
