+++ core/src/main/java/me/wolfyscript/utilities/api/inventory/custom_items/references/APIReference.java
@@ -21,6 +21,10 @@ package me.wolfyscript.utilities.api.inventory.custom_items.references;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.SerializerProvider;
+import me.wolfyscript.utilities.compatibility.plugins.itemsadder.ItemsAdderRef;
+import me.wolfyscript.utilities.compatibility.plugins.mmoitems.MMOItemsRef;
+import me.wolfyscript.utilities.compatibility.plugins.mythicmobs.MythicMobsRef;
+import me.wolfyscript.utilities.compatibility.plugins.oraxen.OraxenRef;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.NotNull;
