+++ plugin-compatibility-module/magic/src/main/java/me/wolfyscript/utilities/compatibility/plugins/magic/MagicRefImpl.java
