+++ plugin-compatibility-module/mmoitems/src/main/java/me/wolfyscript/utilities/compatibility/plugins/mmoitems/MMOItemsRefImpl.java
