+++ plugin-compatibility-module/mythicmobs/src/main/java/me/wolfyscript/utilities/compatibility/plugins/MythicMobsImpl.java
